requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
llama-cpp-python
sentence-transformers
faiss-cpu
numpy
requests
huggingface_hub
